refactor: move loopback redirect helpers to network utils
