+++ platform/backend/src/utils/network.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, test } from "vitest";
-import { isLoopbackAddress } from "./network";
+import {
+  isLoopbackAddress,
+  isLoopbackRedirectUri,
+  loopbackRedirectUriMatchesIgnoringPort,
+} from "./network";
 
 describe("isLoopbackAddress", () => {
   // IPv4 loopback range (127.0.0.0/8)
@@ -51,5 +55,133 @@ describe("isLoopbackAddress", () => {
   test("returns false for invalid input", () => {
     expect(isLoopbackAddress("not-an-ip")).toBe(false);
     expect(isLoopbackAddress("127.0.0")).toBe(false);
+  });
+});
+
+describe("isLoopbackRedirectUri", () => {
+  test("returns true for 127.0.0.1", () => {
+    expect(isLoopbackRedirectUri("http://127.0.0.1:8005/callback")).toBe(true);
+  });
+
+  test("returns true for localhost", () => {
+    expect(isLoopbackRedirectUri("http://localhost:3000/callback")).toBe(true);
+  });
+
+  test("returns true for IPv6 loopback", () => {
+    expect(isLoopbackRedirectUri("http://[::1]:9000/callback")).toBe(true);
+  });
+
+  test("returns true for 127.0.0.1 without port", () => {
+    expect(isLoopbackRedirectUri("http://127.0.0.1/callback")).toBe(true);
+  });
+
+  test("returns false for non-loopback hostname", () => {
+    expect(isLoopbackRedirectUri("https://example.com/callback")).toBe(false);
+  });
+
+  test("returns false for private IP", () => {
+    expect(isLoopbackRedirectUri("http://192.168.1.1/callback")).toBe(false);
+  });
+
+  test("returns false for invalid URI", () => {
+    expect(isLoopbackRedirectUri("not-a-uri")).toBe(false);
+  });
+
+  test("returns false for empty string", () => {
+    expect(isLoopbackRedirectUri("")).toBe(false);
+  });
+});
+
+describe("loopbackRedirectUriMatchesIgnoringPort", () => {
+  test("matches same scheme+host+path with different port", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://127.0.0.1:54321/callback",
+        ["http://127.0.0.1:3000/callback"],
+      ),
+    ).toBe(true);
+  });
+
+  test("matches localhost with different port", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://localhost:54321/callback",
+        ["http://localhost:3000/callback"],
+      ),
+    ).toBe(true);
+  });
+
+  test("matches when requested has port and registered has no port", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://127.0.0.1:54321/callback",
+        ["http://127.0.0.1/callback"],
+      ),
+    ).toBe(true);
+  });
+
+  test("does not match different paths", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://127.0.0.1:54321/other",
+        ["http://127.0.0.1:3000/callback"],
+      ),
+    ).toBe(false);
+  });
+
+  test("does not match different schemes", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "https://127.0.0.1:54321/callback",
+        ["http://127.0.0.1:3000/callback"],
+      ),
+    ).toBe(false);
+  });
+
+  test("does not match localhost vs 127.0.0.1", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://localhost:54321/callback",
+        ["http://127.0.0.1:3000/callback"],
+      ),
+    ).toBe(false);
+  });
+
+  test("does not match non-loopback URI", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "https://example.com:8080/callback",
+        ["https://example.com:3000/callback"],
+      ),
+    ).toBe(false);
+  });
+
+  test("returns false for empty registered URIs", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://127.0.0.1:54321/callback",
+        [],
+      ),
+    ).toBe(false);
+  });
+
+  test("matches against multiple registered URIs", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort(
+        "http://127.0.0.1:54321/callback",
+        [
+          "https://example.com/callback",
+          "http://127.0.0.1:3000/callback",
+        ],
+      ),
+    ).toBe(true);
+  });
+
+  test("returns false for invalid requested URI", () => {
+    expect(
+      loopbackRedirectUriMatchesIgnoringPort("not-a-url", [
+        "http://127.0.0.1:3000/callback",
+      ]),
+    ).toBe(false);
   });
 });
