+++ platform/backend/src/utils/network.ts
@@ -16,3 +16,57 @@ export function isLoopbackAddress(ip: string): boolean {
 
   return isIPv4(ipv4Part) && ipv4Part.startsWith("127.");
 }
+
+/**
+ * Check if a redirect URI targets a loopback address.
+ * Per RFC 8252 Section 7.3, authorization servers must allow any port
+ * for loopback redirect URIs in native app OAuth flows.
+ */
+export function isLoopbackRedirectUri(uri: string): boolean {
+  try {
+    const url = new URL(uri);
+    const hostname = url.hostname.toLowerCase();
+    if (hostname === "localhost") {
+      return true;
+    }
+
+    return isLoopbackAddress(hostname.replace(/^\[(.*)\]$/, "$1"));
+  } catch {
+    return false;
+  }
+}
+
+/**
+ * Check if a requested loopback redirect URI matches any registered URI,
+ * ignoring the port component. Returns true when scheme, host, and path
+ * match but the port differs.
+ */
+export function loopbackRedirectUriMatchesIgnoringPort(
+  requestedUri: string,
+  registeredUris: string[],
+): boolean {
+  if (!isLoopbackRedirectUri(requestedUri)) return false;
+
+  let requestedUrl: URL;
+  try {
+    requestedUrl = new URL(requestedUri);
+  } catch {
+    return false;
+  }
+
+  return registeredUris.some((registeredUri) => {
+    if (!isLoopbackRedirectUri(registeredUri)) return false;
+
+    try {
+      const registeredUrl = new URL(registeredUri);
+      return (
+        requestedUrl.protocol === registeredUrl.protocol &&
+        requestedUrl.hostname.toLowerCase() ===
+          registeredUrl.hostname.toLowerCase() &&
+        requestedUrl.pathname === registeredUrl.pathname
+      );
+    } catch {
+      return false;
+    }
+  });
+}
